Clean up and support for parameterize interfaces
