+++ src/tests/fixtures/generics/genericInterface.ts
