+++ src/tests/fixtures/generics/genericOverArg.ts
@@ -1,0 +1,20 @@
+/** @gqlType */
+export class SomeClass<T> {
+  /** @gqlField */
+  someField(args: { someArg?: T | null }): string {
+    return "someField";
+  }
+}
+
+/** @gqlInput */
+type SomeInput = {
+  someField: string;
+};
+
+/** @gqlType */
+type Query = unknown;
+
+/** @gqlField */
+export function someField(_: Query): SomeClass<SomeInput> {
+  return new SomeClass();
+}
